Fix type conversion portability in MOVSX interpreter
Add intermediate unsigned casts to make sign-extension behavior
well-defined across all platforms. Converting uint64_t directly to
int8_t/int16_t/int32_t is implementation-defined when the value
doesn't fit. Cast through uint8_t/uint16_t/uint32_t first to ensure
portable truncation before sign extension.

+++ vm/ubpf_vm.c
@@ -1007,10 +1007,10 @@ ubpf_exec_ex(
             // MOVSX: sign-extend based on offset value (RFC 9669)
             if (inst.offset == 8) {
                 // Sign-extend 8-bit to 32-bit
-                reg[inst.dst] = (int32_t)(int8_t)reg[inst.src];
+                reg[inst.dst] = (int32_t)(int8_t)(uint8_t)reg[inst.src];
             } else if (inst.offset == 16) {
                 // Sign-extend 16-bit to 32-bit
-                reg[inst.dst] = (int32_t)(int16_t)reg[inst.src];
+                reg[inst.dst] = (int32_t)(int16_t)(uint16_t)reg[inst.src];
             } else {
                 // Normal mov (offset == 0)
                 reg[inst.dst] = reg[inst.src];
@@ -1192,13 +1192,13 @@ ubpf_exec_ex(
             // MOVSX: sign-extend based on offset value (RFC 9669)
             if (inst.offset == 8) {
                 // Sign-extend 8-bit to 64-bit
-                reg[inst.dst] = (int64_t)(int8_t)reg[inst.src];
+                reg[inst.dst] = (int64_t)(int8_t)(uint8_t)reg[inst.src];
             } else if (inst.offset == 16) {
                 // Sign-extend 16-bit to 64-bit
-                reg[inst.dst] = (int64_t)(int16_t)reg[inst.src];
+                reg[inst.dst] = (int64_t)(int16_t)(uint16_t)reg[inst.src];
             } else if (inst.offset == 32) {
                 // Sign-extend 32-bit to 64-bit
-                reg[inst.dst] = (int64_t)(int32_t)reg[inst.src];
+                reg[inst.dst] = (int64_t)(int32_t)(uint32_t)reg[inst.src];
             } else {
                 // Normal mov (offset == 0)
                 reg[inst.dst] = reg[inst.src];
